+++ tnnt_discordbot_cogs/cogs/lookup.py
@@ -45,12 +45,11 @@ from aadiscordbot.cogs.utils.decorators import (
 )
 
 # Terra Nanotech Discordbot Cogs
-from tnnt_discordbot_cogs import __title__
 from tnnt_discordbot_cogs.helper import unload_cog
 from tnnt_discordbot_cogs.models.setting import Setting
-from tnnt_discordbot_cogs.providers import AppLogger
+from tnnt_discordbot_cogs.providers.applogger import AppLogger
 
-logger = AppLogger(my_logger=get_extension_logger(name=__name__), prefix=__title__)
+logger = AppLogger(my_logger=get_extension_logger(name=__name__))
 
 
 class Lookup(commands.Cog):